% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pypi-data.R
\name{all_pypi}
\alias{all_pypi}
\title{Extract all releases and dates for all pypi packages ever}
\usage{
all_pypi(
  chunk_size = 1001,
  results_file = "pypi.Rds",
  data_dir = "./data-temp"
)
}
\arguments{
\item{chunk_size}{Size of chunks into which parallel job is to be broken.
Intermediate results are saved for each chunk.}

\item{results_file}{Name of file (potentially including path) where results
are to be saved.}

\item{data_dir}{Directory in which temporary results for each chunk are to be
saved prior to final aggregation.}
}
\value{
Nothing (data are deposited in 'data_dir`).
}
\description{
Extract all releases and dates for all pypi packages ever
}
\note{
This function may be stopped at any time, and intermediate results
collated with \link{collate_temp_pypi_files}. Passing the name of the file
constructed by that function as \code{results_file} will then re-start analyses
where they left off.
}
