% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pypi-data.R
\name{collate_temp_pypi_files}
\alias{collate_temp_pypi_files}
\title{Collate all temporary files produced by the chunked parallel run of
\code{all_pypi}.}
\usage{
collate_temp_pypi_files(results_file = "pypi.Rds", data_dir = "./data-temp")
}
\arguments{
\item{results_file}{Name of file (potentially including path) where results
are to be saved.}

\item{data_dir}{Directory in which temporary results for each chunk are to be
saved prior to final aggregation.}
}
\value{
Collated results, which are also (re-)saved to specified file
}
\description{
Collate all temporary files produced by the chunked parallel run of
\code{all_pypi}.
}
