% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/couplings.R
\name{coupling_dependencies}
\alias{coupling_dependencies}
\title{Convert raw data (\code{x}) into \code{data.frame} of coupling dependencies as \code{from}
and \code{to} columns for each package.}
\usage{
coupling_dependencies(x)
}
\description{
Convert raw data (\code{x}) into \code{data.frame} of coupling dependencies as \code{from}
and \code{to} columns for each package.
}
