% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/couplings.R
\name{couplings}
\alias{couplings}
\title{Calculate afferent and efferent couplings between packages}
\usage{
couplings(x, year = 2015, summarise = TRUE)
}
\arguments{
\item{x}{Result of \link{load_pkgstats_data}.}

\item{summarise}{if \code{TRUE}, convert full \code{data.frame} of results to summary
statistics.}

\item{y}{Year for which coupling data are to be obtained}
}
\value{
The results
}
\description{
Calculate afferent and efferent couplings between packages
}
