% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/couplings.R
\name{couplings_releases}
\alias{couplings_releases}
\title{Get couplings for each release of each package}
\usage{
couplings_releases(x)
}
\arguments{
\item{x}{Result of \link{load_pkgstats_data}.}
}
\description{
These are couplings between packages but calculated for each release, to
enable examination of changes in coupling stability across releases.
}
