% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran-archiving.R
\name{cran_archiving}
\alias{cran_archiving}
\title{Plot rates of CRAN archiving}
\usage{
cran_archiving(datafile = "pkgstats-results.Rds")
}
\arguments{
\item{datafile}{Name of local file containing CRAN data to load}
}
\value{
A \pkg{ggplot2} object (invisibly)
}
\description{
Plot rates of CRAN archiving
}
