% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{dependencies}
\alias{dependencies}
\title{Extract all dependencies from the "external_calls" component of the main
data.}
\usage{
dependencies(x)
}
\arguments{
\item{x}{Result of \link{load_pkgstats_data}.}
}
\value{
A \code{data.frame} of dependency data with \code{year} column.
}
\description{
Extract all dependencies from the "external_calls" component of the main
data.
}
