% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses-submission-rates.R
\name{pkgstats_analyse_submission_rates}
\alias{pkgstats_analyse_submission_rates}
\title{Analyse submission rates of packages}
\usage{
pkgstats_analyse_submission_rates(datafile)
}
\arguments{
\item{datafile}{Path to local \code{pkgstats} data}
}
\description{
Analyse submission rates of packages
}
