% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgstats-analyses-package.R
\docType{package}
\name{pkgstatsAnalyses-package}
\alias{pkgstatsAnalyses}
\alias{pkgstatsAnalyses-package}
\title{pkgstatsAnalyses: Analyses of 'pkgstats' Results for all CRAN Packages}
\description{
Analyses of 'pkgstats' Results for all CRAN Packages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mpadge/pkgstats-analyses}
  \item Report bugs at \url{https://github.com/mpadge/pkgstats-analyses/issues}
}

}
\author{
\strong{Maintainer}: Mark Padgham \email{mark.padgham@email.com}

}
\keyword{internal}
