% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{recommended_pkgs}
\alias{recommended_pkgs}
\title{Get list of current recommended R packages}
\usage{
recommended_pkgs()
}
\value{
Vector of names of recommended packages
}
\description{
Get list of current recommended R packages
}
