% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/couplings.R
\name{summarise_coupling_data}
\alias{summarise_coupling_data}
\title{Return post-processed summary of coupling data}
\usage{
summarise_coupling_data(x, cran_by_year = TRUE)
}
\arguments{
\item{x}{Result of \link{load_pkgstats_data}.}

\item{cran_by_year}{If \code{TRUE}, implement annual analyses as CRAN would have
existed at each year; otherwise annual analyses are only for packages
uploaded in that year alone.}
}
\value{
A \code{data.frame} of annual summary statistics on coupling instability.
}
\description{
Return post-processed summary of coupling data
}
