% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-data.R
\name{transform_pkgstats_data}
\alias{transform_pkgstats_data}
\title{Transform data to form new variables}
\usage{
transform_pkgstats_data(x)
}
\arguments{
\item{x}{Result of \link{load_pkgstats_data} with \code{raw = TRUE}.}
}
\value{
Transformed version of input, so some variables transformed and new
variables added.
}
\description{
The raw data contain many related raw count variables. This function converts
many of these to relative proportions.
}
