% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-data.R
\name{load_pkgstats_data}
\alias{load_pkgstats_data}
\title{Load one lot of either R or python data}
\usage{
load_pkgstats_data(datafile = "pkgstats-results.Rds")
}
\arguments{
\item{datafile}{Name of local file containing data to load}
}
\value{
The data file, with all dates appropriately converted, and an
additional "month" column added.
}
\description{
Load one lot of either R or python data
}
