% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses-pkg-level.R
\name{pkgstats_analyse_packages}
\alias{pkgstats_analyse_packages}
\title{Analyse individual package trajectories}
\usage{
pkgstats_analyse_packages(x)
}
\arguments{
\item{x}{Result of \link{load_pkgstats_data} and
\link{transform_pkgstats_data}.}
}
\value{
A \code{data.frame} of annual percentage changes in each variable followed
by each package.
}
\description{
Analyse individual package trajectories
}
