% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_r_py}
\alias{plot_r_py}
\title{Time series plot for R versus python}
\usage{
plot_r_py(
  x_r,
  x_p,
  bimonthly = FALSE,
  start_date = "2018-01-01",
  type = "bars"
)
}
\arguments{
\item{x_r}{CRAN data load with \link{load_pkgstats_data} using \code{raw = FALSE}.}

\item{x_p}{pypi data load with \link{load_pkgstats_data} using \code{raw = FALSE}.}

\item{bimonthly}{If \code{TRUE}, aggregate data first into bimonthly intervals
(which generally produces a nicer looking plot).}

\item{start_date}{First date to display, or set to \code{NULL} to display full
range.}

\item{type}{Either "bars" for a bar (column) graph, or "lines" for a line
graph.}
}
\value{
A \pkg{ggplot2} object (invisibly)
}
\description{
Time series plot for R versus python
}
