% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/couplings.R
\name{summarise_coupling_data}
\alias{summarise_coupling_data}
\title{Return post-processed summary of coupling data}
\usage{
summarise_coupling_data(x)
}
\arguments{
\item{x}{Result of \link{load_pkgstats_data}.}
}
\value{
A \code{data.frame} of annual summary statistics on coupling instability.
}
\description{
Return post-processed summary of coupling data
}
